+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 4.2.0 (2014-09-04)
 ------------------
 
+  * Added `stubbles\input\web\WebRequest::id()` which reads the value of an X-Request-ID header or generates a random value in case the header is missing or invalid
   * `stubbles\input\web\WebRequest::uri()` now passes the `stubbles\peer\MalformedUriException` instead of turning it into a `\RuntimeException`